--- a/Gemfile
+++ b/Gemfile
@@ -3,7 +3,7 @@ git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
 ruby "~> 3.2"
 
-gem "rails", "~> 7.0.4"
+gem "rails", "~> 7.1"
 
 gem "active_record_union"
 gem "bcrypt", "~> 3.1.7"