--- a/Gemfile
+++ b/Gemfile
@@ -11,7 +11,7 @@ gem "bootsnap", require: false
 gem "cancancan"
 gem "cssbundling-rails"
 gem "faker", "~> 3.2"
-gem "friendly_id", "~> 5.5.0"
+gem "friendly_id", "~> 5.5"
 gem "humanize", "~> 2.5"
 gem "jbuilder"
 gem "jsbundling-rails"