--- a/Gemfile
+++ b/Gemfile
@@ -52,9 +52,9 @@ end
 group :test do
   gem "capybara"
   gem "rack_session_access"
-  gem "rails-controller-testing", "~> 1.0.5"
+  gem "rails-controller-testing", "~> 1.0"
   gem "selenium-webdriver", "~> 4.7"
-  gem "shoulda-callback-matchers", "~> 1.1.1"
+  gem "shoulda-callback-matchers", "~> 1.1"
   gem "shoulda-matchers", "~> 5.0"
   gem "simplecov", require: false
 end