--- a/app/models/sabeel.rb
+++ b/app/models/sabeel.rb
@@ -6,38 +6,19 @@ class Sabeel < ApplicationRecord
   # * Callbacks
   include NameCallback
 
-  # * Enums
-  # apartment
-  enum :apartment, APARTMENTS
-
   # * FRIENDLY_ID
   include ITSFriendlyId
 
-  # * Methods
-  def address
-    "#{apartment.titleize} #{flat_no}"
-  end
-
-  def taking_thaali?
-    Rails.cache.fetch("sabeel_#{id}_taking_thaali?") { thaalis.exists? year: CURR_YR }
-  end
-
-  def took_thaali?
-    Rails.cache.fetch("sabeel_#{id}_took_thaali?") { thaalis.exists? year: PREV_YR }
-  end
-
-  def last_year_thaali_dues_cleared?
-    thaalis.dues_cleared_in(PREV_YR).present?
-  end
-
   # * RANSACK
-  def self.ransackable_attributes(auth_object = nil)
-    %w[name]
-  end
+  include Ransackable
+  RANSACK_ATTRIBUTES = %w[name]
+  RANSACK_ASSOCIATIONS = %w[thaalis]
 
-  def self.ransackable_associations(auth_object = nil)
-    %w[thaalis]
-  end
+  include ITSValidation
+  include NameValidation
+
+  # * Enums
+  enum :apartment, APARTMENTS
 
   # * Scopes
   scope :no_thaali, -> { where.missing(:thaalis) }
@@ -57,16 +38,25 @@ class Sabeel < ApplicationRecord
   scope :with_thaali_size, ->(size) { thaalis.where(thaalis: {size:}) }
 
   # * Validations
-  # apartment
   validates :apartment, presence: true
-  # Email
   validates_email_format_of :email, allow_blank: true
-  # Flat No
   validates :flat_no, numericality: {only_integer: true, greater_than: 0}
-  # ITS
-  include ITSValidation
-  # name
-  include NameValidation
-  # mobile
   validates :mobile, numericality: {only_integer: true}, length: {is: 10}
+
+  # * Methods
+  def address
+    "#{apartment.titleize} #{flat_no}"
+  end
+
+  def taking_thaali?
+    Rails.cache.fetch("sabeel_#{id}_taking_thaali?") { thaalis.exists? year: CURR_YR }
+  end
+
+  def took_thaali?
+    Rails.cache.fetch("sabeel_#{id}_took_thaali?") { thaalis.exists? year: PREV_YR }
+  end
+
+  def last_year_thaali_dues_cleared?
+    thaalis.dues_cleared_in(PREV_YR).present?
+  end
 end