--- a/app/models/transaction.rb
+++ b/app/models/transaction.rb
@@ -1,15 +1,12 @@
 class Transaction < ApplicationRecord
   default_scope { order(date: :desc) }
 
-  # * RANSACK
-  include Ransackable
-  RANSACK_ATTRIBUTES = %w[recipe_no]
-
   # * Associations
   belongs_to :thaali
 
-  # * Enums
-  enum :mode, MODES
+  # * RANSACK
+  include Ransackable
+  RANSACK_ATTRIBUTES = %w[recipe_no]
 
   # * FRIENDLY_ID
   extend FriendlyId
@@ -19,18 +16,17 @@ class Transaction < ApplicationRecord
     recipe_no_changed?
   end
 
+  # * Enums
+  enum :mode, MODES
+
   # * Scopes
   scope :that_occured_on, ->(date) { where(date: date) }
 
   # * Validations
-  # amount
   validates :amount, :recipe_no, numericality: {only_integer: true, greater_than: 0}
   validate :amount_to_be_less_than_balance, if: :will_save_change_to_amount?
-  # date
   validates_date :date, on_or_before: :today, if: :will_save_change_to_date?
-  # date & mode
   validates :mode, :date, presence: true
-  # recipe no
   validates :recipe_no, uniqueness: true
 
   private