--- a/app/views/sabeels/index.html.erb
+++ b/app/views/sabeels/index.html.erb
@@ -3,7 +3,7 @@
 
 <%# * search %>
 <%= render partial: "shared/search",
-      locals: {message: "Name or ITS Nnumber.", path: sabeels_path, search: :name_or_its_cont} %>
+      locals: {attr_name: "Name", path: sabeels_path, search: :name_cont} %>
 
 <%# * show all sabeels %>
 <%= render partial: "shared/results",