--- a/app/views/shared/_search.html.erb
+++ b/app/views/shared/_search.html.erb
@@ -1,7 +1,7 @@
 <div id="search" class="col-md-4 offset-md-4 mt-5" role="search">
   <%= search_form_for @q, url: path, data: {controller: "search", debounce_target: "form", turbo_frame: :results} do |f| %>
 
-    <% f.search_field search, placeholder: "Search By #{message}", class: "form-control", type: "search", autocomplete: "off", data: {action: "input->search#submit"},
+    <% f.search_field search, placeholder: "Search by #{attr_name}", class: "form-control", type: "search", autocomplete: "off", data: {action: "input->search#submit"},
          # disables search with enter keypress
          onkeypress: "return event.keyCode != 13;" %>
   <% end %>