--- a/config/application.rb
+++ b/config/application.rb
@@ -22,10 +22,15 @@ module Fmb
   class Application < Rails::Application
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 7.0
-    config.exceptions_app = routes
+
     config.time_zone = "Asia/Kolkata"
     config.i18n.available_locales = [:en]
     config.i18n.default_locale = :en
+
+    # Please, add to the `ignore` list any other `lib` subdirectories that do
+    # not contain `.rb` files, or that should not be reloaded or eager loaded.
+    # Common ones are `templates`, `generators`, or `middleware`, for example.
+    config.autoload_lib(ignore: %w[assets tasks])
 
     # Configuration for the application, engines, and railties goes here.
     #