--- a/config/initializers/assets.rb
+++ b/config/initializers/assets.rb
@@ -5,7 +5,7 @@ Rails.application.config.assets.version = "1.0"
 
 # Add additional assets to the asset load path.
 # Rails.application.config.assets.paths << Emoji.images_path
-Rails.application.config.assets.paths << Rails.root.join("node_modules/bootstrap-icons/font")
+# Rails.application.config.assets.paths << Rails.root.join("node_modules/bootstrap-icons/font")
 
 # Precompile additional assets.
 # application.js, application.css, and all non-JS/CSS in the app/assets