… .ruby-version
… Gemfile.lock
… app/models/concerns/i_t_s_friendly_id.rb
… config/initializers/new_framework_defaults_7_1.rb
… spec/features/sabeels/index_spec.rb
… spec/features/thaalis/index_spec.rb
… spec/features/transactions/index_spec.rb
… spec/features/transactions/new_spec.rb
… spec/support/selenium_setup.rb