This PR includes the following changes:

## Gems

| Gem | Before | Now |
|-----|------:|-----:|
| bundler | 2.3.26 | 2.4.21 |
| bcrypt | 3.1.18 | 3.1.19 |
| capybara | 3.38.0 | 3.39.2 |
| faker | 3.1.0 | 3.2.2 |
| friendly_id| 5.5.0 | 5.5.1 |
| pagy | 6.0 | 6.2 |
| pg | 1.4.5 | 1.5.4 |
| puma | 5.6.5 | 6.4.0 | 
| rails | 7.0.4.3 | 7.1.2 |
| ransack | 3.2.1 | 4.1.1 |
| redis | 4.0 | 5.0.8 |
| ruby | 3.1.2 | 3.2.2 |
| selenium-webdriver | 4.7 | 4.15 |
| simple_form | 5.1 | 5.3 |

----

## Tests

Updating the `capybara` & `selenium-webdriver` gems have significantly **decreased the runtime** of tests.

| Metric    | Before  |   Now   | Percentage Improvement |
|-----------|--------:|--------:|--------------:|
| Load time |     ~1.6 secs    |     ~1.2secs    | **25%** |
| Run time  | ~48 sec | ~40 sec |  **16.67%**  |

-----

## CI
We can see similar performance improvement in the runtime of GitHub Action (lint-n-test):

| PR  | Before  |   After   | Percentage Improvement |
|-----------|--------:|--------:|--------------:|
| [Rails Way](https://github.com/JuzerShakir/fmb/pull/26)  | ~4min 40 sec |  |  |
| [Update Fly.io & CI](https://github.com/JuzerShakir/fmb/pull/27)  |  | ~2min 40 sec |  **42.9%**  |
| This PR | ~2min 40 sec | ~1min 30secs |  **43.75%**  |

Overall improvement is **68.75%** from _~4mins 40secs_ to _~1min 30secs_.